gsl1

+++ src/async_wrap.cc
@@ -613,12 +613,12 @@ void AsyncWrap::AsyncReset(double execution_async_id, bool silent) {
           TRACING_CATEGORY_NODE1(async_hooks))) {                             \
         auto data = tracing::TracedValue::Create();                           \
         data->SetInteger("executionAsyncId",                                  \
-                         static_cast<int64_t>(env()->execution_async_id()));  \
+                         gsl::narrow<int>(env()->execution_async_id()));  \
         data->SetInteger("triggerAsyncId",                                    \
-                         static_cast<int64_t>(get_trigger_async_id()));       \
+                         gsl::narrow<int>(get_trigger_async_id()));       \
         TRACE_EVENT_NESTABLE_ASYNC_BEGIN1(                                    \
           TRACING_CATEGORY_NODE1(async_hooks),                                \
-          #PROVIDER, static_cast<int64_t>(get_async_id()),                    \
+          #PROVIDER, gsl::narrow<int>(get_async_id()),                    \
           "data", std::move(data));                                           \
         }                                                                     \
       break;
